- update delete method

+++ src/main/java/by/itech/lab/supplier/service/impl/WarehouseServiceImpl.java
@@ -127,15 +127,12 @@ public class WarehouseServiceImpl implements WarehouseService {
     @Transactional
     @Override
     public void deleteByIds(final List<Long> id) {
-//        System.out.println(applicationService.isExistNotFinishedStatus(id) + " app");
-//        System.out.println(itemInWarehouseRepository.findWarehouseItemIdsByWarehouseIds(id).isEmpty() + " item");
-
-//        if (applicationService.isExistNotFinishedStatus(id) &&
-//                itemInWarehouseRepository.findWarehouseItemIdsByWarehouseIds(id).isEmpty()){
+        if (applicationService.isWarehousesFreeFromApplications(id) &&
+                itemInWarehouseRepository.findWarehouseItemIdsByWarehouseIds(id).isEmpty()){
 
             warehouseRepository.deleteByIds(id);
             userService.deleteWarehousesForAllUsers(id);
-//        }
+        }
     }
 
     @Transactional
